Refactors as separate components
