--- resources/assets/components/utilities/SchoolFinder/school-finder.scss
@@ -1,7 +1,0 @@
-@import '../../../scss/next-toolbox';
-
-.school-finder {
-  > .card.rounded {
-    overflow: visible;
-  }
-}
